# openapi.yaml
openapi: 3.1.0
info:
  title: Thordata Firecrawl API
  description: Turn any website into AI-ready data with a single API
  version: 0.1.0
paths:
  /health:
    get:
      summary: Health
      description: Health check endpoint.
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /:
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
  /playground:
    get:
      summary: Playground
      operationId: playground_playground_get
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
  /v1/scrape:
    post:
      summary: Scrape Endpoint
      description: Scrape a single URL.
      operationId: scrape_endpoint_v1_scrape_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrapeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrapeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/batch-scrape:
    post:
      summary: Batch Scrape Endpoint
      description: Batch scrape multiple URLs (Firecrawl-style batch endpoint).
      operationId: batch_scrape_endpoint_v1_batch_scrape_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScrapeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchScrapeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crawl:
    post:
      summary: Crawl Submit
      description: Submit an async crawl job. Use GET /v1/crawl/{id} to poll status/results.
      operationId: crawl_submit_v1_crawl_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrawlRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlJobResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crawl/{job_id}:
    get:
      summary: Crawl Status
      description: Get crawl job status and results (supports pagination).
      operationId: crawl_status_v1_crawl__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/crawl/{job_id}/cancel:
    post:
      summary: Crawl Cancel
      description: Cancel a crawl job (best-effort).
      operationId: crawl_cancel_v1_crawl__job_id__cancel_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/map:
    post:
      summary: Map Endpoint
      description: Discover URLs on a website.
      operationId: map_endpoint_v1_map_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/search:
    post:
      summary: Search Endpoint
      description: Search the web.
      operationId: search_endpoint_v1_search_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/search-and-scrape:
    post:
      summary: Search And Scrape Endpoint
      description: 'Combined search + scrape helper: search the web, then scrape the
        top results.'
      operationId: search_and_scrape_endpoint_v1_search_and_scrape_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAndScrapeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAndScrapeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/agent:
    post:
      summary: Agent Endpoint
      description: Run an agent task for structured extraction.
      operationId: agent_endpoint_v1_agent_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AgentRequest:
      properties:
        prompt:
          type: string
          title: Prompt
        urls:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Urls
        schema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Schema
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        searchLimit:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Searchlimit
          default: 3
        formats:
          items:
            type: string
          type: array
          title: Formats
          description: Formats to scrape for context
          default:
          - markdown
        scrapeOptions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scrapeoptions
      type: object
      required:
      - prompt
      title: AgentRequest
    AgentResponse:
      properties:
        success:
          type: boolean
          title: Success
        data:
          additionalProperties: true
          type: object
          title: Data
        sources:
          items:
            type: string
          type: array
          title: Sources
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - success
      - data
      - sources
      title: AgentResponse
    BatchScrapeRequest:
      properties:
        urls:
          items:
            type: string
          type: array
          title: Urls
        formats:
          items:
            type: string
          type: array
          title: Formats
          description: Output formats
          default:
          - markdown
        scrapeOptions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scrapeoptions
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - urls
      title: BatchScrapeRequest
    BatchScrapeResponse:
      properties:
        success:
          type: boolean
          title: Success
        results:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Results
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - success
      - results
      title: BatchScrapeResponse
    CrawlJobResponse:
      properties:
        success:
          type: boolean
          title: Success
        id:
          type: string
          title: Id
        url:
          type: string
          title: Url
      type: object
      required:
      - success
      - id
      - url
      title: CrawlJobResponse
    CrawlRequest:
      properties:
        url:
          type: string
          title: Url
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 100
        scrapeOptions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scrapeoptions
        includeSubdomains:
          type: boolean
          title: Includesubdomains
          default: false
        maxDepth:
          anyOf:
          - type: integer
            maximum: 10.0
            minimum: 1.0
          - type: 'null'
          title: Maxdepth
      type: object
      required:
      - url
      title: CrawlRequest
    CrawlStatusResponse:
      properties:
        status:
          type: string
          title: Status
        total:
          type: integer
          title: Total
        completed:
          type: integer
          title: Completed
        failed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Failed
          default: 0
        data:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Data
      type: object
      required:
      - status
      - total
      - completed
      - data
      title: CrawlStatusResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MapRequest:
      properties:
        url:
          type: string
          title: Url
        search:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      type: object
      required:
      - url
      title: MapRequest
    MapResponse:
      properties:
        success:
          type: boolean
          title: Success
        links:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Links
      type: object
      required:
      - success
      - links
      title: MapResponse
    ScrapeRequest:
      properties:
        url:
          type: string
          title: Url
        formats:
          items:
            type: string
          type: array
          title: Formats
          description: Output formats
          default:
          - markdown
        scrapeOptions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scrapeoptions
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - url
      title: ScrapeRequest
    ScrapeResponse:
      properties:
        success:
          type: boolean
          title: Success
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - success
      title: ScrapeResponse
    SearchAndScrapeRequest:
      properties:
        query:
          type: string
          title: Query
        searchLimit:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Searchlimit
          default: 5
        formats:
          items:
            type: string
          type: array
          title: Formats
          description: Output formats for scraping
          default:
          - markdown
        scrapeOptions:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scrapeoptions
      type: object
      required:
      - query
      title: SearchAndScrapeRequest
    SearchAndScrapeResponse:
      properties:
        success:
          type: boolean
          title: Success
        query:
          type: string
          title: Query
        results:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Results
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - success
      - query
      - results
      title: SearchAndScrapeResponse
    SearchRequest:
      properties:
        query:
          type: string
          title: Query
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          default: 10
        engine:
          anyOf:
          - type: string
          - type: 'null'
          title: Engine
          default: google
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
      type: object
      required:
      - query
      title: SearchRequest
    SearchResponse:
      properties:
        success:
          type: boolean
          title: Success
        data:
          additionalProperties: true
          type: object
          title: Data
      type: object
      required:
      - success
      - data
      title: SearchResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError